rename userHelper.js to userHelpers.js
